Add area of affect to projectiles

+++ Assets/Scripts/Building/Towers/Cannon/CannonSettings.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Cannon Settings", menuName = "Tower/Cannon/Settings")]
 public class CannonSettings : TowerBaseSettings
 {
-    public int Area;
+    public float Area;
 }
